extract messaging from game_state module
